rename wevotedsection to wewaitedsection
